requirements.txt
Flask==2.1.1
Flask-PyMongo==2.3.0
PyJWT==2.1.0
pymongo==3.12.0
bcrypt==3.2.0
pymongo[srv]
